backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1

pymongo==4.6.1
motor==3.3.2

# ML (Stable for Render + Training)
tensorflow-cpu==2.10.0
h5py==3.7.0
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.3.2
scipy==1.10.1
joblib==1.3.2

# Image + OCR
pillow==10.3.0
opencv-python-headless==4.9.0.80
pytesseract==0.3.13
pdf2image==1.17.0
openpyxl==3.1.5

# Auth
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.5.0
Authlib==1.6.9
email-validator==2.3.0

# HTTP
requests==2.32.5
httpx==0.28.1

# Core utils
itsdangerous==2.2.0
python-multipart==0.0.22

# AI
groq==1.1.1
requests